+++ changelog.md
@@ -14,6 +14,10 @@
 
 * [4090](https://github.com/zeta-chain/node/pull/4090) - print error message in detail if unable to decode Bitcoin memo
 
+### Tests
+
+* [4071](https://github.com/zeta-chain/node/pull/4071) - use v2 connector contract in admin e2e tests
+
 ## v33.0.0
 
 ### Features